調整LV1的景物材質參數

+++ Assets/HelloPico2/Data/VFX/GGDog/Formal/Model_Shader/Toon_CubeMap/Uber_ToonShader.shader
@@ -60,7 +60,7 @@ Shader "GGDog/Uber_ToonShader"
 			{
 				half4 uv : TEXCOORD0;
 				half4 vertex : SV_POSITION;
-                float3 normal_VS : TEXCOORD1;
+                half3 normal_VS : TEXCOORD1;
 			};
 			
             sampler2D _MainTex;
